Lift test:e2e to monorepo root.
It tests website, which depends on a running api, so it spans projects.
Later it may be desirable to move some of /playwright.config.ts into /packages/website/ (and other specific packages).
