+++ .github/workflows/website.yml
@@ -49,6 +49,9 @@ jobs:
       - run: npx playwright install --with-deps ${{ matrix.browser }} chromium
         working-directory: packages/website
       - run: xvfb-run --auto-servernum --server-args="-screen 0 1280x960x24" -- npm -w @web3-storage/website run test:e2e -- --project=${{ matrix.browser }}
+        env:
+          NEXT_PUBLIC_API: http://localhost:8787
+          NEXT_PUBLIC_ENV: dev
       - name: Check for website test results
         id: check_website_test_results
         uses: andstor/file-existence-action@v1
